requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.0
greenlet==1.1.2
idna==3.3
numpy==1.23.1
pandas==1.4.3
psycopg2==2.9.3
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.39
SQLAlchemy-Utils==0.38.2
urllib3==1.26.10
